pyproject.toml
[project]
name = "researchagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-core>=1.2.16",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "nest-asyncio>=1.6.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.54.0",
    "tavily-python>=0.7.22",
]
